+++ AlphaFS/Filesystem/File Class/File.CopyMove.cs
@@ -1198,7 +1198,7 @@ namespace Alphaleonis.Win32.Filesystem
 
                      if (!isFolder)
                      {
-                        using (SafeFileHandle safeHandle = CreateFileCore(transaction, sourceFileNameLp, ExtendedFileAttributes.None, null, FileMode.Open, 0, FileShare.Read, false, PathFormat.LongFullPath))
+                        using (SafeFileHandle safeHandle = CreateFileCore(transaction, sourceFileNameLp, ExtendedFileAttributes.Normal, null, FileMode.Open, 0, FileShare.Read, false, PathFormat.LongFullPath))
                            if (safeHandle != null && safeHandle.IsInvalid)
                               fileNameLp = sourceFileNameLp;
                      }